move tera-operator as well
